refactor: ObserverBottom 커스텀훅으로 분리 #448
