+++ frontend-admin/src/pages/ContentPage/hooks/useObserverBottom.ts
@@ -1,0 +1,18 @@
+import { useRef, useState } from "react";
+
+const useObserverBottom = () => {
+  const [isBottomVisible, setIsBottomVisible] = useState(false);
+  const observerRef = useRef<HTMLDivElement | null>(null);
+
+  const handleReachBottom = () => {
+    setIsBottomVisible(true);
+  };
+
+  const handleLeaveBottom = () => {
+    setIsBottomVisible(false);
+  };
+
+  return { isBottomVisible, handleReachBottom, handleLeaveBottom, observerRef };
+};
+
+export default useObserverBottom;
